split ModPresetRow
